refactor: introduce AxisManager
