+++ svg-time-series/src/chart/data.ts
@@ -11,18 +11,6 @@ export interface IMinMax {
   readonly max: number;
 }
 
-function buildMinMax(fst: Readonly<IMinMax>, snd: Readonly<IMinMax>): IMinMax {
-  return {
-    min: Math.min(fst.min, snd.min),
-    max: Math.max(fst.max, snd.max),
-  } as const;
-}
-
-const minMaxIdentity: IMinMax = {
-  min: Infinity,
-  max: -Infinity,
-};
-
 export interface IDataSource {
   readonly startTime: number;
   readonly timeStep: number;
@@ -140,30 +128,6 @@ export class ChartData {
 
   private clampIndex(idx: number): number {
     return Math.min(Math.max(idx, 0), this.data.length - 1);
-  }
-
-  private buildAxisMinMax(axis: number): Array<IMinMax | undefined> {
-    const idxs = this.seriesByAxis[axis];
-    return this.data.map((row) => {
-      let min = Infinity;
-      let max = -Infinity;
-      for (const j of idxs) {
-        const val = row[j];
-        if (Number.isFinite(val)) {
-          if (val < min) min = val;
-          if (val > max) max = val;
-        }
-      }
-      return min !== Infinity ? ({ min, max } as IMinMax) : undefined;
-    });
-  }
-
-  buildAxisTree(axis: number): SegmentTree<IMinMax> {
-    const arr = Array.from(
-      this.buildAxisMinMax(axis),
-      (v) => v ?? minMaxIdentity,
-    );
-    return new SegmentTree(arr, buildMinMax, minMaxIdentity);
   }
 
   bAxisVisible(bIndexVisible: AR1Basis, tree: SegmentTree<IMinMax>): AR1Basis {
